requirements.txt
anndata==0.7.8
blitzgsea @ git+https://github.com/MaayanLab/blitzgsea.git@9752dd7c5f12d4c935cfeb944f3b045dc23ac2fb
biopython==1.79
bioservices==1.8.4
gseapy==0.10.8
h5py==3.6.0
ipykernel==6.7.0
matplotlib==3.5.1
matplotlib-venn==0.11.6
missingpy==0.2.0
networkx==2.7.1
numpy==1.21.5
pandas==1.4.1
PyYAML==6.0
scanpy==1.8.2
scikit-learn==1.0.2
scikit-misc==0.1.3
scipy==1.7.3
seaborn==0.11.2
statsmodels==0.13.2
supervenn==0.4.1
torch==1.8.0
torch-scatter==2.0.9
torch-sparse==0.6.12
tqdm==4.63.0
umap-learn==0.5.2
wandb==0.12.9
